Split CoreMotion use on Slip Timer
